100% front end

+++ quizs/src/main/java/com/quizly/quizs/repositories/ParticipantRepository.java
@@ -4,5 +4,5 @@ import com.quizly.quizs.models.Participant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ParticipantRepository extends JpaRepository<Participant, Long> {
-    int countParticipantByNameOfParticipant(String participant);
+    Long countParticipantByNameOfParticipant(String participant);
 }
